+++ Changelog.md
@@ -13,6 +13,7 @@
 - Fixes some config false positives error messages.
 - Fixes command line arguments parser handling of `--` for switching to verbatim mode (#670).
 - Fixes rendering of U+E0B2  in pixel-perfect box drawing mode (#707).
+- Fixes rendering of cursor sometimes being almost invisible when foreground and background colors are similar/equal (#691).
 - Changes `XTSMGRAPHICS` to match implementation of xterm *exactly* when querying sixel image limits, to be capped at terminal viewport dimensions (#656).
 - Win32: Attaches to parent console if present, so typing `contour help` in a terminal actually shows something.
 - Implements UTF-8 encoded mouse transport (`CSI ? 1005 h`)